test/usb_driver/requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
kiwisolver==1.4.8
matplotlib>=3.5.0
numpy>=1.20.0
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
pip==25.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyparsing==3.2.1
pyserial>=3.5
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
setuptools==75.8.0
six==1.17.0
wheel==0.45.1
